feat: implement idempotency key handling for message and room creation - Added `IdempotencyHeader` type to encapsulate idempotency key in GraphQL context. - Updated GraphQL handlers to utilize the `Idempotency-Key` header. - Enhanced message and room services to check idempotency cache before creating new entries. - Added tracing. WIP.

+++ libs/service/src/mappers/room_mapper.rs
@@ -9,7 +9,7 @@ impl EntityToDomain<DomainRoom> for EntityRoom {
         DomainRoom {
             id: self.public_id,
             name: self.name,
-            capacity: self.capacity,
+            capacity: self.capacity as u32,
             description: self.description,
         }
     }
